Messaging_Service/messaging-service/requirements.txt
sendgrid
twilio
flask
functions_framework
google-cloud-secret-manager